SERVER: Fixed error:
The function acceptPEers was the function that was allocating memory for
the array of peerAddresses. Whenever the function returned, the pointer
pointing to the newly allocated memory was not translated back to main.
So, when the pointer was used to call createRing, the pointer was NULL.
To fix the problem the memory was allocated in main instead.