PEER: Beginning to implement the exit functionality.
The peers will only be able to exit when they have the token. They will
send a message out advertizing that they what to exit with their port
number and the port number of the their next peer.
The exiting peer will wait for the message to get back before exiting.
There will be a problem if a peer decides to join a peer that is
exiting. To solve this problem, need to go back and readjust
getNextPeerFromPeer to send the port number of the joining peer and not
a message from the user.
Becuase of this, processNextMessage will have to be modified to handle
this because the function assumes the sender is the one that is joining
the ring. However, if the exiting peer forwards the joining peer's port
number, the recieving peer will think that the exiting peer is joining.